Library/requirements.txt
Django==1.5.1
coverage==3.6
dj-database-url==0.2.2
dj-static==0.0.5
django-dajaxice==0.5.5
django-jquery==1.9.1
django-openid-auth==0.5
django-recaptcha==0.0.6
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
pysolr==3.1.0
python-openid==2.2.5
requests==1.2.3
selenium==2.35.0
static==0.4
wsgiref==0.1.2
django-crispy-forms==1.4.0
django-pure-pagination==0.2.1
PIL==1.1.6